Remove core dependency on viz
