+++ viz/pkg/api_utils_test.go
@@ -1,4 +1,4 @@
-package util
+package pkg
 
 import (
 	"errors"
