+++ viz/tap/api/grpc_server.go
@@ -12,7 +12,6 @@ import (
 
 	httpPb "github.com/linkerd/linkerd2-proxy-api/go/http_types"
 	proxy "github.com/linkerd/linkerd2-proxy-api/go/tap"
-	apiUtil "github.com/linkerd/linkerd2/controller/api/util"
 	netPb "github.com/linkerd/linkerd2/controller/gen/common/net"
 	"github.com/linkerd/linkerd2/controller/k8s"
 	"github.com/linkerd/linkerd2/pkg/addr"
@@ -20,6 +19,7 @@ import (
 	"github.com/linkerd/linkerd2/pkg/prometheus"
 	"github.com/linkerd/linkerd2/pkg/util"
 	metricsPb "github.com/linkerd/linkerd2/viz/metrics-api/gen/viz"
+	"github.com/linkerd/linkerd2/viz/pkg"
 	vizLabels "github.com/linkerd/linkerd2/viz/pkg/labels"
 	tapPb "github.com/linkerd/linkerd2/viz/tap/gen/tap"
 	log "github.com/sirupsen/logrus"
@@ -75,7 +75,7 @@ func (s *GRPCTapServer) TapByResource(req *tapPb.TapByResourceRequest, stream ta
 
 	objects, err := s.k8sAPI.GetObjects(res.GetNamespace(), res.GetType(), res.GetName(), labelSelector)
 	if err != nil {
-		return apiUtil.GRPCError(err)
+		return pkg.GRPCError(err)
 	}
 
 	pods := []*corev1.Pod{}
@@ -84,7 +84,7 @@ func (s *GRPCTapServer) TapByResource(req *tapPb.TapByResourceRequest, stream ta
 	for _, object := range objects {
 		podsFor, err := s.k8sAPI.GetPodsFor(object, false)
 		if err != nil {
-			return apiUtil.GRPCError(err)
+			return pkg.GRPCError(err)
 		}
 
 		for _, pod := range podsFor {
@@ -132,7 +132,7 @@ func (s *GRPCTapServer) TapByResource(req *tapPb.TapByResourceRequest, stream ta
 
 	match, err := makeByResourceMatch(req.GetMatch())
 	if err != nil {
-		return apiUtil.GRPCError(err)
+		return pkg.GRPCError(err)
 	}
 
 	extract := &proxy.ObserveRequest_Extract{}
@@ -169,7 +169,7 @@ func (s *GRPCTapServer) TapByResource(req *tapPb.TapByResourceRequest, stream ta
 		case event := <-events:
 			err := stream.Send(event)
 			if err != nil {
-				return apiUtil.GRPCError(err)
+				return pkg.GRPCError(err)
 			}
 		}
 	}
